refactor: entity 패키지명 domain으로 수정 (#15)
